+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 
 ## Version 2025.2
 
+### Next release
+
+* 🔨 Import and Auto Backup: do not crash if reading file name fails.
+
 ### 2025.2.3 - 2025-06-26 🧪
 
 * 🔧 Backup: rename to "Export" and "Import", move them and Auto Backup up to More screen.